Move run_yara() to scanners package
